requirements.txt
python-dotenv
Flask
azure-ai-textanalytics==5.3.0